# Conflicts:
#	src/main/java/sg/edu/nus/comp/cs4218/impl/util/FileSystemUtils.java

+++ src/main/java/sg/edu/nus/comp/cs4218/impl/util/FileSystemUtils.java
@@ -7,6 +7,12 @@ public final class FileSystemUtils {
 
     private FileSystemUtils() {}
 
+    /**
+     * Creates a file in the current working directory
+     *
+     * @param filename  Name of file
+     */
+
     public static boolean fileOrDirExist(String filename) {
         return new File(filename).exists();
     }
@@ -28,7 +34,16 @@ public final class FileSystemUtils {
             throw new FileOrDirDoesNotExistException(filename);
         }
 
+    /**
+     * Deletes a file in the current working directory
+     *
+     * @param filename  Name of file
+     */
+    public static void deleteFile(String filename) throws FileDoesNotExistException, FileDeletionException {
         File file = new File(filename);
+        if (!file.exists()) {
+            throw new FileDoesNotExistException(filename);
+        }
 
         if (!file.delete()) {
             throw new FileOrDirDeletionException(filename);
